Code quality and performance changes
