+++ packages/eslint-plugin/src/rules/no-empty-styled-expression/__tests__/rule.test.ts
@@ -20,6 +20,13 @@ tester.run('no-styled-tagged-template-expression', noStyledEmptyExpressionRule, 
         color: blue
       });
     `,
+    `
+      import { styled } from 'styled';
+
+      styled.span({
+        color: blue
+      });
+    `,
   ],
   invalid: createInvalidTestCases([
     {
