+++ packages/eslint-plugin/src/rules/no-styled-tagged-template-expression/utils.ts
@@ -1,17 +1,9 @@
 import type { Rule, Scope } from 'eslint';
 
-import { COMPILED_IMPORT } from '../../utils/constants';
+import { isStyledImportSpecifier } from '../../utils/styled-import';
 
-type Definition = Scope.Definition;
 type Node = Rule.Node;
 type Reference = Scope.Reference;
-
-const isStyledImportSpecifier = (def: Definition) =>
-  def.node.type === 'ImportSpecifier' &&
-  def.node.imported.type === 'Identifier' &&
-  def.node.imported.name === 'styled' &&
-  def.parent?.type === 'ImportDeclaration' &&
-  def.parent?.source.value === COMPILED_IMPORT;
 
 export const isStyled = (node: Node, references: Reference[]): boolean =>
   (node.type === 'MemberExpression' &&
