+++ packages/eslint-plugin/src/utils/styled-import.ts
@@ -1,0 +1,11 @@
+import type { Scope } from 'eslint';
+
+import { COMPILED_IMPORT } from './constants';
+type Definition = Scope.Definition;
+
+export const isStyledImportSpecifier = (def: Definition): boolean =>
+  def.node.type === 'ImportSpecifier' &&
+  def.node.imported.type === 'Identifier' &&
+  def.node.imported.name === 'styled' &&
+  def.parent?.type === 'ImportDeclaration' &&
+  def.parent?.source.value === COMPILED_IMPORT;
